// src/types/supabase.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      feedback_categories: {
        Row: {
          created_at: string | null
          created_by: string | null
          created_by_type: string | null
          description: string | null
          icon: string | null
          id: string
          is_archived: boolean | null
          is_default: boolean | null
          name: string
          post_count: number | null
          slug: string
          updated_at: string | null
        }
        Insert: {
          created_at?: string | null
          created_by?: string | null
          created_by_type?: string | null
          description?: string | null
          icon?: string | null
          id?: string
          is_archived?: boolean | null
          is_default?: boolean | null
          name: string
          post_count?: number | null
          slug: string
          updated_at?: string | null
        }
        Update: {
          created_at?: string | null
          created_by?: string | null
          created_by_type?: string | null
          description?: string | null
          icon?: string | null
          id?: string
          is_archived?: boolean | null
          is_default?: boolean | null
          name?: string
          post_count?: number | null
          slug?: string
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "feedback_categories_created_by_fkey"
            columns: ["created_by"]
            isOneToOne: false
            referencedRelation: "pending_pilot_registrations"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "feedback_categories_created_by_fkey"
            columns: ["created_by"]
            isOneToOne: false
            referencedRelation: "pilot_users"
            referencedColumns: ["id"]
          },
        ]
      }
      feedback_comments: {
        Row: {
          author_display_name: string
          author_rank: string | null
          content: string
          created_at: string | null
          edited_at: string | null
          flag_count: number | null
          id: string
          is_anonymous: boolean | null
          is_edited: boolean | null
          is_flagged: boolean | null
          parent_comment_id: string | null
          pilot_user_id: string | null
          post_id: string
          updated_at: string | null
        }
        Insert: {
          author_display_name: string
          author_rank?: string | null
          content: string
          created_at?: string | null
          edited_at?: string | null
          flag_count?: number | null
          id?: string
          is_anonymous?: boolean | null
          is_edited?: boolean | null
          is_flagged?: boolean | null
          parent_comment_id?: string | null
          pilot_user_id?: string | null
          post_id: string
          updated_at?: string | null
        }
        Update: {
          author_display_name?: string
          author_rank?: string | null
          content?: string
          created_at?: string | null
          edited_at?: string | null
          flag_count?: number | null
          id?: string
          is_anonymous?: boolean | null
          is_edited?: boolean | null
          is_flagged?: boolean | null
          parent_comment_id?: string | null
          pilot_user_id?: string | null
          post_id?: string
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "feedback_comments_parent_comment_id_fkey"
            columns: ["parent_comment_id"]
            isOneToOne: false
            referencedRelation: "feedback_comments"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "feedback_comments_pilot_user_id_fkey"
            columns: ["pilot_user_id"]
            isOneToOne: false
            referencedRelation: "pending_pilot_registrations"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "feedback_comments_pilot_user_id_fkey"
            columns: ["pilot_user_id"]
            isOneToOne: false
            referencedRelation: "pilot_users"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "feedback_comments_post_id_fkey"
            columns: ["post_id"]
            isOneToOne: false
            referencedRelation: "feedback_posts"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "feedback_comments_post_id_fkey"
            columns: ["post_id"]
            isOneToOne: false
            referencedRelation: "feedback_posts_feed"
            referencedColumns: ["id"]
          },
        ]
      }
      feedback_posts: {
        Row: {
          author_display_name: string
          author_rank: string | null
          category_id: string | null
          comment_count: number | null
          content: string
          created_at: string | null
          flag_count: number | null
          id: string
          is_anonymous: boolean | null
          is_pinned: boolean | null
          last_activity_at: string | null
          pilot_user_id: string | null
          pin_order: number | null
          status: string | null
          tags: string[] | null
          title: string
          updated_at: string | null
          view_count: number | null
        }
        Insert: {
          author_display_name: string
          author_rank?: string | null
          category_id?: string | null
          comment_count?: number | null
          content: string
          created_at?: string | null
          flag_count?: number | null
          id?: string
          is_anonymous?: boolean | null
          is_pinned?: boolean | null
          last_activity_at?: string | null
          pilot_user_id?: string | null
          pin_order?: number | null
          status?: string | null
          tags?: string[] | null
          title: string
          updated_at?: string | null
          view_count?: number | null
        }
        Update: {
          author_display_name?: string
          author_rank?: string | null
          category_id?: string | null
          comment_count?: number | null
          content?: string
          created_at?: string | null
          flag_count?: number | null
          id?: string
          is_anonymous?: boolean | null
          is_pinned?: boolean | null
          last_activity_at?: string | null
          pilot_user_id?: string | null
          pin_order?: number | null
          status?: string | null
          tags?: string[] | null
          title?: string
          updated_at?: string | null
          view_count?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "feedback_posts_category_id_fkey"
            columns: ["category_id"]
            isOneToOne: false
            referencedRelation: "feedback_categories"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "feedback_posts_category_id_fkey"
            columns: ["category_id"]
            isOneToOne: false
            referencedRelation: "feedback_posts_feed"
            referencedColumns: ["category_id"]
          },
          {
            foreignKeyName: "feedback_posts_pilot_user_id_fkey"
            columns: ["pilot_user_id"]
            isOneToOne: false
            referencedRelation: "pending_pilot_registrations"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "feedback_posts_pilot_user_id_fkey"
            columns: ["pilot_user_id"]
            isOneToOne: false
            referencedRelation: "pilot_users"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}